refactor: :truck: move github package resolution into its own directory
