+++ src/packages/github/index.ts
@@ -1,0 +1,3 @@
+export { GithubPackage, type GithubPackageReleaseData } from "./package";
+export { GithubPackageResolver } from "./resolver";
+export { GithubPackageVersion } from "./version";
